2023 Google I/O Connect

+++ diet/src/main/java/com/example/diet/Domain/Request.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Request {
-    private Long req_id;
+    private Integer req_id;
     private String from_usr_id;
     private String to_usr_id;    //一开始是phone   需要从phone改成usr_id
     private String req_msg;    //关系
